refactor: clean up message decoding and encoding
